pyproject.toml
[project]
name = "youtube-downloader"
version = "0.1.0"
description = "Youtube Downloader microservice"
requires-python = ">=3.10"
dependencies = [
    "anthropic==0.39.0",
    "python-dotenv==1.0.1",
    "typing-extensions==4.12.2",
    "redis==5.2.0",
    "minio==7.2.10",
    "openai==1.54.4",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["src/tests"]
asyncio_mode = "auto"
asyncio_fixture_loop_scope = "function"
